requirements.txt
flask
pytube
openai
python-dotenv
gunicorn